kvm-bindings, kvm-ioctls, vmm-sys-util crate upgrade
kvm-ioctl v0.12.0 introduces some broken api changes and expose some new api.
There are some code in the Repositories that overlap with the new API functions,
so upgrade to newest kvm-ioctl crate.

+++ cpu/src/aarch64/mod.rs
@@ -193,7 +193,7 @@ impl ArmCPUState {
             .with_context(|| "Failed to init kvm vcpu")?;
         self.mpidr = vcpu_fd
             .get_one_reg(SYS_MPIDR_EL1)
-            .with_context(|| "Failed to get mpidr")?;
+            .with_context(|| "Failed to get mpidr")? as u64;
 
         ArmCPUState::set_cpu_feature(vcpu_config, vcpu_fd)?;
         self.features = *vcpu_config;
